backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30

sqlalchemy>=2.0
psycopg[binary]>=3.2

alembic>=1.13

passlib[bcrypt]>=1.7.4

pydantic-settings>=2.2
python-dotenv>=1.0

